Create post with testing for h2

+++ rev-connect-api/src/main/java/com/rev_connect_api/repositories/PostRepository.java
@@ -8,9 +8,9 @@ import java.math.BigInteger;
 import java.util.Optional;
 
 @Repository
-public interface PostRepository extends JpaRepository<Post, Integer> {
+public interface PostRepository extends JpaRepository<Post, BigInteger> {
 
     Optional<Post> getPostByPostId(BigInteger id);
-
     void deletePostByPostId(BigInteger id);
+    
 }
